LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/manual/imagebitmap/createImageBitmap-colorSpaceConversion-expected.txt


FAIL createImageBitmap from a bitmap HTMLImageElement, and drawImage on the created ImageBitmap with colorSpaceConversion: none assert_approx_equals: Red channel of the pixel at (5, 5) expected 124 +/- 10 but got 255
FAIL createImageBitmap from a Blob, and drawImage on the created ImageBitmap with colorSpaceConversion: none assert_approx_equals: Red channel of the pixel at (5, 5) expected 124 +/- 10 but got 255
